backend/requirements.txt
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
msgpack==1.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
redis==5.2.0
sqlparse==0.5.1
